Growable frame stack and register array (#596)
* Growable frame stack and register array (#593)

Replace fixed-size frame and register arrays with heap-allocated slices
that grow on demand. Non-tail-recursive list functions like filter and
enumerate-interval previously overflowed at ~475 elements due to the
480-frame hard limit. Now they handle 5000+ elements, bounded only by
available memory (hard caps: 32768 frames, 65536 registers).

- VM struct: frames/registers changed from [N]T to []T, allocated from
  gc.allocator in init(), freed in deinit()
- ensureFrameCapacity/ensureRegisterCapacity: double-on-overflow growth
- All frame push sites (4) and register checks (~15) updated to use
  dynamic bounds or growth calls
- Continuations: heap-allocate saved_frames temp, grow-on-restore
- Fibers: dynamic backing arrays, save/restore handles size mismatches
- GC: fiber deallocation frees backing arrays, size tracking updated


* Update dev docs for growable frame/register stacks

- deep-recursion-register-overflow.md: mark fully resolved, update
  constants and fix descriptions to reflect growable slices
- gc-reachability-bug.md: note deep force/promise overflow is resolved
- architecture.md: VM description mentions growable stacks


* Widen CallFrame.base from u16 to u32

The u16 base field overflowed at ~20,000 non-tail frames, causing a Zig
panic instead of a clean StackOverflow error. Widening to u32 supports
up to 4 billion register indices, far beyond the 65536 hard cap.

Only base is widened — dst stays u16 since it's a per-frame offset
bounded by locals_count (~255 max). The struct stays 48 bytes on 64-bit
(the u32 fills former alignment padding).

Also updates robustness tests: the "deep non-tail recursion" test now
verifies growable stack success at 10000 frames and clean StackOverflow
at 50000 frames (past the 32768 frame hard cap).


* Update deep-recursion postmortem for u32 base widening

Add fix #4 (u16→u32 widening) to the fixes-applied section and update
the code example to show u32.


* Fix error-format test for growable stack

The stack overflow error format test used (deep 10000) which now
succeeds with growable stacks. Increase to 50000 to exceed the 32768
frame hard cap and trigger a clean StackOverflow error.


---------

+++ src/types.zig
@@ -351,7 +351,7 @@ pub const SavedFrame = struct {
     native: ?*NativeFn,
     code: []const u8,
     ip: usize,
-    base: u16,
+    base: u32,
     dst: u16,
     saved_wind_count: u16,
     // On wasm32, pointer-sized fields shrink from 8 to 4 bytes, making the
@@ -383,7 +383,7 @@ pub const Continuation = struct {
     wind_records: []WindRecord,
     wind_count: usize,
     dst_reg: u16, // register offset within frame where result goes
-    dst_base: u16, // base register of the return frame
+    dst_base: u32, // base register of the return frame
     // Single backing allocation holding registers, frames, handlers and winds
     // contiguously. The four slices above are views into this buffer; it is
     // freed as one block on sweep. Empty for escape continuations.
